+++ src/onegov.election_day/onegov/election_day/tests/views/test_views_election.py
@@ -140,26 +140,21 @@ def test_view_election_party_strengths(election_day_app_gr):
     assert '/election/proporz-election/party-strengths-data' in chart
 
     export = client.get('/election/proporz-election/data-parties').text
-    assert export == (
-        'year,name,id,total_votes,color,mandates,votes,'
-        'panachage_votes_from_0,panachage_votes_from_1,panachage_votes_from_2'
-        '\r\n'
-        '2015,BDP,0,11270,#0571b0,1,60387,,,\r\n'
-        '2015,CVP,1,11270,#0571b0,1,49117,,,\r\n'
-        '2015,FDP,2,11270,#0571b0,0,35134,,,\r\n'
-    )
+    lines = export.split('\r\n')
+    assert lines[0].startswith('year,name,id,total_votes,color,mandates,votes')
+    assert lines[1].startswith('2015,BDP,0,11270,#0571b0,1,60387')
+    assert lines[2].startswith('2015,CVP,1,11270,#0571b0,1,49117')
+    assert lines[3].startswith('2015,FDP,2,11270,#0571b0,0,35134')
 
     # Historical data
     csv_parties = (
-        'year,name,id,total_votes,color,mandates,votes,'
-        'panachage_votes_from_0,panachage_votes_from_1,panachage_votes_from_2'
-        '\r\n'
-        '2015,BDP,0,60000,#efb52c,1,10000,,,\r\n'
-        '2015,CVP,1,60000,#ff6300,1,30000,,,\r\n'
-        '2015,FDP,2,60000,#4068c8,0,20000,,,\r\n'
-        '2011,BDP,0,40000,#efb52c,1,1000,,,\r\n'
-        '2011,CVP,1,40000,#ff6300,1,15000,,,\r\n'
-        '2011,FDP,2,40000,#4068c8,1,10000,,,\r\n'
+        'year,name,id,total_votes,color,mandates,votes\r\n'
+        '2015,BDP,0,60000,#efb52c,1,10000\r\n'
+        '2015,CVP,1,60000,#ff6300,1,30000\r\n'
+        '2015,FDP,2,60000,#4068c8,0,20000\r\n'
+        '2011,BDP,0,40000,#efb52c,1,1000\r\n'
+        '2011,CVP,1,40000,#ff6300,1,15000\r\n'
+        '2011,FDP,2,40000,#4068c8,1,10000\r\n'
     ).encode('utf-8')
 
     upload = client.get('/election/proporz-election/upload-party-results')
@@ -221,7 +216,11 @@ def test_view_election_party_strengths(election_day_app_gr):
     assert '8.3%' in results
 
     export = client.get('/election/proporz-election/data-parties').text
-    assert export.encode('utf-8') == csv_parties
+    lines = export.split('\r\n')
+    lines_csv = csv_parties.decode('utf-8').split('\r\n')
+    assert all([
+        line.startswith(lines_csv[index]) for index, line in enumerate(lines)
+    ])
 
 
 def test_view_election_connections(election_day_app_gr):