+++ src/onegov.election_day/onegov/election_day/utils/d3_renderer.py
@@ -13,6 +13,8 @@ from onegov.election_day.views.election.connections \
 from onegov.election_day.views.election.lists import view_election_lists_data
 from onegov.election_day.views.election.lists_panachage import \
     view_election_lists_panachage_data
+from onegov.election_day.views.election.parties_panachage import \
+    view_election_parties_panachage_data
 from onegov.election_day.views.election.party_strengths \
     import view_election_party_strengths_data
 from onegov.election_day.views.election_compound.party_strengths \
@@ -162,11 +164,20 @@ class D3Renderer():
                 chart = self.get_chart('grouped', fmt, data)
         return (chart, data) if return_data else chart
 
-    def get_panachage_chart(self, item, fmt, return_data=False):
+    def get_lists_panachage_chart(self, item, fmt, return_data=False):
         chart = None
         data = None
         if isinstance(item, Election):
             data = view_election_lists_panachage_data(item, None)
+            if data and data.get('links') and data.get('nodes'):
+                return self.get_chart('sankey', fmt, data)
+        return (chart, data) if return_data else chart
+
+    def get_parties_panachage_chart(self, item, fmt, return_data=False):
+        chart = None
+        data = None
+        if isinstance(item, Election):
+            data = view_election_parties_panachage_data(item, None)
             if data and data.get('links') and data.get('nodes'):
                 return self.get_chart('sankey', fmt, data)
         return (chart, data) if return_data else chart