+++ src/onegov.election_day/onegov/election_day/utils/svg_generator.py
@@ -57,9 +57,11 @@ class SvgGenerator():
         if type_ == 'map':
             chart = self.renderer.get_map_chart(item, 'svg', locale)
         if type_ == 'lists-panachage':
-            chart = self.renderer.get_panachage_chart(item, 'svg')
+            chart = self.renderer.get_lists_panachage_chart(item, 'svg')
         if type_ == 'party-strengths':
             chart = self.renderer.get_party_strengths_chart(item, 'svg')
+        if type_ == 'parties-panachage':
+            chart = self.renderer.get_parties_panachage_chart(item, 'svg')
         if chart:
             with self.app.filestorage.open(path, 'w') as f:
                 copyfileobj(chart, f)
@@ -88,6 +90,7 @@ class SvgGenerator():
                 self.generate_svg(election, 'lists')
                 self.generate_svg(election, 'connections')
                 self.generate_svg(election, 'party-strengths')
+                self.generate_svg(election, 'parties-panachage')
                 self.generate_svg(election, 'lists-panachage')
         for election_compound in self.session.query(ElectionCompound):
             self.generate_svg(election_compound, 'party-strengths')